layout/style/test/test_bug1089417.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1089417
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1089417</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1089417 **/

  SimpleTest.waitForExplicitFinish();

  function run() {
    var f = document.getElementById("f");
    var fwin = f.contentWindow;
    var fdoc = f.contentDocument;

    f.height = "400";
    fdoc.getElementById("s").disabled = false;
    is(fwin.getComputedStyle(fdoc.documentElement).backgroundColor,
       "rgb(0, 128, 0)",
       "media query change should have restyled");

    f.height = "200";
    fdoc.getElementById("s").disabled = true;
    fdoc.getElementById("s").disabled = false;
    is(fwin.getComputedStyle(fdoc.documentElement).backgroundColor,
       "rgb(255, 0, 0)",
       "media query change should have restyled");
    SimpleTest.finish();
  }

  </script>
</head>
<body onload="run()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1089417">Mozilla Bug 1089417</a>
<div id="display">
  <iframe id="f" src="file_bug1089417_iframe.html" width="300" height="200"></iframe>
</div>
<pre id="test">
</pre>
</body>
</html>
